Rename public Untrusted/Trusted paths to host/enclave
For consistency with rest of Open Enclave SDK
