+++ new_platforms/include/optee/enclave/Simulator/compiler.h
