+++ new_platforms/include/optee/enclave/Simulator/corecrt_wtime.h
