+++ new_platforms/include/optee/enclave/Simulator/fcntl.h
