+++ new_platforms/include/optee/enclave/Simulator/oeenclave.h
