+++ new_platforms/include/optee/enclave/Simulator/riot_cyres.h
