+++ new_platforms/include/optee/enclave/Simulator/sys/stat.h
