+++ new_platforms/include/optee/enclave/errno.h
