+++ new_platforms/include/optee/enclave/mbusafecrt.h
