+++ new_platforms/include/optee/enclave/sgx_edger8r.h
