+++ new_platforms/include/optee/enclave/sgx_lfence.h
