+++ new_platforms/include/optee/enclave/sgxoveroptee_enclave.h
