+++ new_platforms/include/optee/host/Simulator/compiler.h
