+++ new_platforms/include/optee/host/Simulator/oehost.h
