+++ new_platforms/include/optee/host/sgx_edger8r.h
