+++ new_platforms/include/sgx/enclave/intrin.h
