+++ new_platforms/libsocket/enc/optee/sub.mk
@@ -9,7 +9,7 @@ ROOT_RELATIVE_PATH = ../../../
 
 CFLAGS += -DOE_USE_OPTEE
 
-global-incdirs-y += $(ROOT_RELATIVE_PATH)include/optee/Trusted
+global-incdirs-y += $(ROOT_RELATIVE_PATH)include/optee/enclave
 global-incdirs-y += $(ROOT_RELATIVE_PATH)include/optee
 global-incdirs-y += $(ROOT_RELATIVE_PATH)include
 global-incdirs-y += $(OpteeDir)lib/libutee/include
