+++ new_platforms/libstdio/enc/oestdio_enc.vcxproj
@@ -259,7 +259,7 @@
       <RuntimeLibrary>MultiThreadedDebug</RuntimeLibrary>
       <Optimization>Disabled</Optimization>
       <WarningLevel>Level3</WarningLevel>
-      <AdditionalIncludeDirectories>.;$(NewPlatformsDir)include\optee\Trusted\Simulator;$(NewPlatformsDir)include\optee\Trusted;$(NewPlatformsDir)include\optee;$(NewPlatformsDir)include;$(OESdkDir)include;$(SGXSDKInstallPath)include;$(OESdkDir)3rdparty\optee_os\lib\libutee\include;$(OESdkDir)3rdparty\optee_os\lib\libutils\ext\include;%(AdditionalIncludeDirectories)</AdditionalIncludeDirectories>
+      <AdditionalIncludeDirectories>.;$(NewPlatformsDir)include\optee\enclave\Simulator;$(NewPlatformsDir)include\optee\enclave;$(NewPlatformsDir)include\optee;$(NewPlatformsDir)include;$(OESdkDir)include;$(SGXSDKInstallPath)include;$(OESdkDir)3rdparty\optee_os\lib\libutee\include;$(OESdkDir)3rdparty\optee_os\lib\libutils\ext\include;%(AdditionalIncludeDirectories)</AdditionalIncludeDirectories>
       <PreprocessorDefinitions>OE_SIMULATE_OPTEE;</PreprocessorDefinitions>
     </ClCompile>
     <Lib>
@@ -285,7 +285,7 @@
       <RuntimeLibrary>MultiThreadedDebug</RuntimeLibrary>
       <Optimization>Disabled</Optimization>
       <WarningLevel>Level3</WarningLevel>
-      <AdditionalIncludeDirectories>.;$(NewPlatformsDir)include\optee\Trusted\Simulator;$(NewPlatformsDir)include\optee\Trusted;$(NewPlatformsDir)include\optee;$(NewPlatformsDir)include;$(OESdkDir)include;$(SGXSDKInstallPath)include;$(OESdkDir)3rdparty\optee_os\lib\libutee\include;$(OESdkDir)3rdparty\optee_os\lib\libutils\ext\include;%(AdditionalIncludeDirectories)</AdditionalIncludeDirectories>
+      <AdditionalIncludeDirectories>.;$(NewPlatformsDir)include\optee\enclave\Simulator;$(NewPlatformsDir)include\optee\enclave;$(NewPlatformsDir)include\optee;$(NewPlatformsDir)include;$(OESdkDir)include;$(SGXSDKInstallPath)include;$(OESdkDir)3rdparty\optee_os\lib\libutee\include;$(OESdkDir)3rdparty\optee_os\lib\libutils\ext\include;%(AdditionalIncludeDirectories)</AdditionalIncludeDirectories>
       <PreprocessorDefinitions>OE_SIMULATE_OPTEE;_ARM_;</PreprocessorDefinitions>
     </ClCompile>
     <Lib>
@@ -389,7 +389,7 @@
       <RuntimeLibrary>MultiThreadedDebug</RuntimeLibrary>
       <Optimization>Disabled</Optimization>
       <WarningLevel>Level3</WarningLevel>
-      <AdditionalIncludeDirectories>.;$(NewPlatformsDir)include\optee\Trusted\Simulator;$(NewPlatformsDir)include\optee\Trusted;$(NewPlatformsDir)include\optee;$(NewPlatformsDir)include;$(OESdkDir)include;$(SGXSDKInstallPath)include;$(OESdkDir)3rdparty\optee_os\lib\libutee\include;$(OESdkDir)3rdparty\optee_os\lib\libutils\ext\include;%(AdditionalIncludeDirectories)</AdditionalIncludeDirectories>
+      <AdditionalIncludeDirectories>.;$(NewPlatformsDir)include\optee\enclave\Simulator;$(NewPlatformsDir)include\optee\enclave;$(NewPlatformsDir)include\optee;$(NewPlatformsDir)include;$(OESdkDir)include;$(SGXSDKInstallPath)include;$(OESdkDir)3rdparty\optee_os\lib\libutee\include;$(OESdkDir)3rdparty\optee_os\lib\libutils\ext\include;%(AdditionalIncludeDirectories)</AdditionalIncludeDirectories>
       <PreprocessorDefinitions>OE_SIMULATE_OPTEE;</PreprocessorDefinitions>
     </ClCompile>
     <Lib>
