+++ new_platforms/samples/sockets/enclave/SampleTA.c
