+++ new_platforms/samples/sockets/enclave/SampleTA.config.xml
