+++ new_platforms/samples/sockets/enclave/SampleTA.vcxproj.filters
