+++ new_platforms/samples/sockets/enclave/SampleTA_private.pem
