+++ new_platforms/samples/sockets/enclave/optee/linux_gcc.mak
