+++ new_platforms/samples/sockets/enclave/optee/sub.mk
