+++ new_platforms/samples/sockets/enclave/optee/user_ta_header_defines.h
