+++ new_platforms/samples/sockets/enclave/optee/uuids.reg
