+++ new_platforms/samples/sockets/host/SampleClientApp/Dockerfile
