+++ new_platforms/samples/sockets/host/SampleClientApp/SampleClientApp.vcxproj.filters
