+++ new_platforms/samples/sockets/host/SampleClientApp/build_container.sh
