+++ new_platforms/samples/sockets/host/SampleClientApp/main.c
