+++ new_platforms/samples/sockets/host/SampleServerApp/SampleServerApp.vcxproj.filters
