+++ new_platforms/samples/sockets/host/SampleServerApp/main.c
